finished a myriad of parts, will need to go over them again to make sure i didn't miss anything. genitals added, but i know for a fact i'm going to need a lot of functions in it i havent added. added an interface so body parts can subscribe and be notified to changes in femininity

+++ BodyPartsFinished/SpecialInteraction/IGrowShrinkable.cs
@@ -12,9 +12,9 @@ namespace CoC.BodyParts.SpecialInteraction
 	interface IGrowShrinkable
 	{
 		bool CanReducto();
-		int UseReducto();
+		float UseReducto();
 		bool CanGrowPlus();
-		int UseGroPlus();
+		float UseGroPlus();
 
 	}
 }
